requirements.txt
Flask
redis
mysql-connector-python
unittest2